pyproject.toml
[tool.poetry]
name = "djangorest"
version = "0.1.0"
description = ""
authors = ["Edgar Valdez <edgarvaldez99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
Markdown = "^3.5"
django-filter = "^23.3"
Django = "^4.2.6"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
drf-extra-fields = "^3.7.0"
Pillow = "^10.0.1"

[tool.poetry.dev-dependencies]
black = "^23.9.1"
mypy = "^1.6.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
nox = "^2023.4.22"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = "manage.py"
ignore_missing_imports = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
